split subdivision out of smoothing
